Try to test Java enums with Scala.js
